moving docs folder inside utils package
